backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15
django-cors-headers>=4.3
django-jazzmin>=3.0
pandas>=2.2
openpyxl>=3.1
